Refactor command handling
